Remove token value from API response and Mitre UI
Removed also IamLogoutSuccessHandler and related tests
Minor fixes
