+++ iam-login-service/src/main/java/it/infn/mw/iam/config/security/IamWebSecurityConfig.java
@@ -79,7 +79,6 @@ import it.infn.mw.iam.config.mfa.IamTotpMfaProperties;
 import it.infn.mw.iam.core.IamLocalAuthenticationProvider;
 import it.infn.mw.iam.core.oauth.TokenEndpointJwtClientAuthFilter;
 import it.infn.mw.iam.core.oidc.AuthorizationRequestFilter;
-import it.infn.mw.iam.core.web.loginpage.IamLogoutSuccessHandler;
 import it.infn.mw.iam.persistence.repository.IamAccountRepository;
 import it.infn.mw.iam.persistence.repository.IamX509CertificateRepository;
 import it.infn.mw.iam.service.aup.AUPSignatureCheckService;
@@ -141,9 +140,6 @@ public class IamWebSecurityConfig {
 
     @Autowired
     private IamTotpMfaProperties iamTotpMfaProperties;
-
-    @Autowired
-    private IamLogoutSuccessHandler iamLogoutSuccessHandler;
 
     @Autowired
     private Clock clock;
@@ -214,7 +210,6 @@ public class IamWebSecurityConfig {
           .addFilterAfter(extendedHttpServletRequestFilter(), UsernamePasswordAuthenticationFilter.class)
         .logout()
           .logoutUrl("/logout")
-          .logoutSuccessHandler(iamLogoutSuccessHandler)
         .and().anonymous()
         .and()
           .csrf()
