+++ iam-login-service/src/main/java/it/infn/mw/iam/core/oauth/profile/IamRegistrationTokenService.java
@@ -57,7 +57,6 @@ import it.infn.mw.iam.persistence.repository.IamOAuthAccessTokenRepository;
 
 @SuppressWarnings("deprecation")
 @Service
-@Primary
 @Transactional
 public class IamRegistrationTokenService implements RegistrationTokenService {
